requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.5.1
dacite==1.6.0
-e .
